Mobile view updated. Not complete yet.
Made the navigation bar responsive so it can have a hamburger menu on mobile. Edited all the elements so there was no overspill creating that yucky white space that lets you scroll to the side on mobile. Got the "main" section looking nice on mobile. Not sure i'm 100% happy with the main section yet.